👩‍💻 Reorganize old export commands as part of pull
