+++ .changeset/lemon-starfishes-jam.md
@@ -1,0 +1,5 @@
+---
+"@curvenote/cli": patch
+---
+
+Reorganize old export commands as part of pull
