--- packages/curvenote-cli/src/export/index.ts
@@ -1,4 +1,0 @@
-export * from './markdown/index.js';
-export * from './notebook/index.js';
-export * from './jupyter-book/index.js';
-export * from './utils/index.js';
