+++ packages/curvenote-cli/src/frontmatter.ts
@@ -11,8 +11,9 @@ import {
 import type { ValidationOptions } from 'simple-validators';
 import { filterKeys } from 'simple-validators';
 import { dirname, join } from 'node:path';
-import type { ISession } from '../session/index.js';
-import { THUMBNAILS_FOLDER } from '../utils/index.js';
+import type { ISession } from './session/index.js';
+
+export const THUMBNAILS_FOLDER = 'thumbnails';
 
 export function saveAffiliations(session: ISession, project: Project) {
   session.store.dispatch(
