+++ packages/curvenote-cli/src/index.ts
@@ -1,12 +1,11 @@
+export * from './export.js';
 export * from './session/index.js';
 export * from './models.js';
-export * from './export/index.js';
 export * from './utils/index.js';
-export * as web from './web/index.js';
-export * as exp from './export/index.js';
-export * as sync from './sync/index.js';
 export * as check from './check/index.js';
-export * as works from './works/index.js';
 export * as submissions from './submissions/index.js';
+export * as sync from './sync/index.js';
 export * as upload from './uploads/index.js';
+export * as web from './web/index.js';
+export * as works from './works/index.js';
 export { default as version } from './version.js';
