+++ packages/curvenote-cli/src/submissions/types.ts
@@ -1,6 +1,8 @@
 import type { CollectionDTO, SubmissionKindDTO } from '@curvenote/common';
 import type { CompiledCheckResults } from '../check/types.js';
 
+export type STATUS_ACTIONS = 'publish' | 'unpublish';
+
 export type SubmitOpts = {
   kind?: string;
   collection?: string;
@@ -63,3 +65,28 @@ export type NewCheckJobResults = {
     report?: CompiledCheckResults;
   };
 };
+
+export interface CreateSubmissionBody {
+  work_version_id: string;
+  collection_id: string;
+  kind_id: string;
+  draft: boolean;
+  job_id: string;
+}
+
+export interface UpdateSubmissionBody {
+  work_version_id: string;
+  job_id: string;
+}
+
+export interface CreateCliCheckJobPostBody {
+  job_type: 'CLI_CHECK';
+  payload: Record<string, any>;
+  results: Record<string, any>;
+}
+
+export interface UpdateCliCheckJobPostBody {
+  status: string;
+  message: string;
+  results: Record<string, any>;
+}
