+++ packages/curvenote-cli/src/sync/index.ts
@@ -1,3 +1,3 @@
 export { init } from './init.js';
-export { pull } from './pull.js';
+export { pull } from './pull/project.js';
 export { clone } from './clone.js';
