+++ packages/curvenote-cli/src/sync/pull/jupyter-book/index.ts
@@ -1,0 +1,1 @@
+export { projectToJupyterBook } from './project.js';
