+++ packages/curvenote-cli/src/sync/pull/utils/walkArticle.ts
@@ -12,12 +12,12 @@ import {
 } from '@curvenote/blocks';
 import type { Nodes } from '@curvenote/schema';
 import { DEFAULT_IMAGE_WIDTH, nodeNames, ReferenceKind } from '@curvenote/schema';
-import { Block, Version } from '../../models.js';
-import type { ISession } from '../../session/types.js';
+import { Block, Version } from '../../../models.js';
+import type { ISession } from '../../../session/types.js';
 import { basekey } from './basekey.js';
 import { getEditorState, getEditorStateFromHTML } from './getEditorState.js';
 import { getImageSrc } from './getImageSrc.js';
-import { getBlockAndLatestVersion } from './getLatest.js';
+import { getBlockAndLatestVersion } from './getBlockAndLatestVersion.js';
 
 export interface ArticleStateChild {
   state: ReturnType<typeof getEditorState>;
