+++ packages/curvenote-cli/src/utils/index.ts
@@ -1,2 +1,2 @@
 export * from './utils.js';
-export * from './types.js';
+export * from './logVersions.js';
